Work on my ECS input system

+++ ECS/SpaceInvadersMyECS/Assets/Source/Systems/SystemEnemySetup.cs
@@ -15,17 +15,16 @@ public class SystemEnemySetup : ISystem
 
     public void update()
     {
+        var enemyEntities = EntityManager.getEntities(enemiesToSetupFilter);
+        if(enemyEntities.Count == 0) return;
         var entityTarget = EntityManager.getEntities(targetComponentsFilter)[0];
         var targetPosition = EntityManager.GetComponent<ComponentsTransform>(EntityManager.getComponents(entityTarget)).transform;
 
-        var enemyEntities = EntityManager.getEntities(enemiesToSetupFilter);
-        Debug.Log(enemyEntities.Count);
         for (int i = 0; i < enemyEntities.Count; i++)
         {
             var components = EntityManager.getComponents(enemyEntities[i]);
             var transform = EntityManager.GetComponent<ComponentsTransform>(components);
             var setup = EntityManager.GetComponent<ComponentEnemySetup>(components);
-            Debug.Log(setup.position.ToString("F2"));
             transform.transform.position = setup.position;
             transform.transform.LookAt(targetPosition);
             EntityManager.removeComponent(enemyEntities[i], setup);
